Big updates.. lot more jsr107 function..
Added multiple examples of jsr107 usage via annotations, 
and one via non annotated. Need to revisit the modifications made to the
annotations layer though, it needs to use a service model to locate the
initial cache, rather than having that vcap parse (and more critically
the Redisson clinet instance) out where the app code can't get to it.

As part of the jsr107 examples, implemented a basic item framework,
allowing items to be added to rooms with minimal coding required, and
have item commands etc be managed by the room.

Flipped RoomEndpoint to use a cdi bean to manage the connected sessions,
and use that elsewhere to transmit responses, which decouples the
endpoint from all the places that previously needed session+endpoint to
send replies, plus allows for things to be sent by the room without a
request (eg, toggle state change is broadcast to the room)